Added a single test for localaddress , handled review comments
